Disjoining operations from the BigInteger and BigDecimal classes (for readability)
