+++ src/Deveel.Math/Deveel.Math.csproj
@@ -157,7 +157,14 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Deveel.Math\BigDecimal.cs" />
+    <Compile Include="Deveel.Math\BigDecimal_Convertible.cs" />
+    <Compile Include="Deveel.Math\BigDecimal_Functions.cs" />
+    <Compile Include="Deveel.Math\BigDecimal_Operators.cs" />
+    <Compile Include="Deveel.Math\BigDecimal_Parse.cs" />
     <Compile Include="Deveel.Math\BigInteger.cs" />
+    <Compile Include="Deveel.Math\BigInteger_Convertible.cs" />
+    <Compile Include="Deveel.Math\BigInteger_Functions.cs" />
+    <Compile Include="Deveel.Math\BitInteger_Operators.cs" />
     <Compile Include="Deveel.Math\BitLevel.cs" />
     <Compile Include="Deveel.Math\CharHelper.cs" />
     <Compile Include="Deveel.Math\Conversion.cs" />
