+++ src/Deveel.Math/Deveel.Math/BigDecimal.cs
@@ -33,9 +33,9 @@ namespace Deveel.Math {
 	[Serializable]
 #endif
 	[System.Diagnostics.DebuggerDisplay("{ToString()}")]
-	public sealed class BigDecimal : IComparable<BigDecimal>, IEquatable<BigDecimal>
+	public sealed partial class BigDecimal : IComparable<BigDecimal>, IEquatable<BigDecimal>
 #if !PORTABLE
-		, ISerializable, IConvertible
+		, ISerializable
 #endif
 	{
 
@@ -291,7 +291,7 @@ namespace Deveel.Math {
 				if (_bitLength < 64) {
 					smallValue = mantisa << (-_scale);
 				} else {
-					intVal = BigInteger.ValueOf(mantisa).ShiftLeft(-_scale);
+					intVal = BigInteger.FromInt64(mantisa).ShiftLeft(-_scale);
 				}
 				_scale = 0;
 			} else if (_scale > 0) {
@@ -300,7 +300,7 @@ namespace Deveel.Math {
 					smallValue = mantisa*LongFivePow[_scale];
 					_bitLength = BitLength(smallValue);
 				} else {
-					SetUnscaledValue(Multiplication.MultiplyByFivePow(BigInteger.ValueOf(mantisa), _scale));
+					SetUnscaledValue(Multiplication.MultiplyByFivePow(BigInteger.FromInt64(mantisa), _scale));
 				}
 			} else {
 				// scale == 0
@@ -700,12 +700,12 @@ namespace Deveel.Math {
 			var largerSignum = larger.Sign;
 			if (largerSignum == smaller.Sign) {
 				tempBi = Multiplication.MultiplyByPositiveInt(larger.GetUnscaledValue(), 10)
-					.Add(BigInteger.ValueOf(largerSignum));
+					.Add(BigInteger.FromInt64(largerSignum));
 			} else {
 				tempBi = larger.GetUnscaledValue().Subtract(
-					BigInteger.ValueOf(largerSignum));
+					BigInteger.FromInt64(largerSignum));
 				tempBi = Multiplication.MultiplyByPositiveInt(tempBi, 10)
-					.Add(BigInteger.ValueOf(largerSignum*9));
+					.Add(BigInteger.FromInt64(largerSignum*9));
 			}
 			// Rounding the improved adding 
 			larger = new BigDecimal(tempBi, larger._scale + 1);
@@ -812,11 +812,11 @@ namespace Deveel.Math {
 					BigInteger tempBI;
 					if (thisSignum != subtrahend.Sign) {
 						tempBI = Multiplication.MultiplyByPositiveInt(this.GetUnscaledValue(), 10)
-							.Add(BigInteger.ValueOf(thisSignum));
+							.Add(BigInteger.FromInt64(thisSignum));
 					} else {
-						tempBI = GetUnscaledValue().Subtract(BigInteger.ValueOf(thisSignum));
+						tempBI = GetUnscaledValue().Subtract(BigInteger.FromInt64(thisSignum));
 						tempBI = Multiplication.MultiplyByPositiveInt(tempBI, 10)
-							.Add(BigInteger.ValueOf(thisSignum*9));
+							.Add(BigInteger.FromInt64(thisSignum*9));
 					}
 					// Rounding the improved subtracting
 					var leftOperand = new BigDecimal(tempBI, _scale + 1); // it will be only the left operand (this) 
@@ -940,974 +940,6 @@ namespace Deveel.Math {
 		#endregion
 
 		#region Operations
-
-		private static BigDecimal DivideBigIntegers(BigInteger scaledDividend, BigInteger scaledDivisor, int scale,
-			RoundingMode roundingMode) {
-			BigInteger remainder;
-			BigInteger quotient = scaledDividend.DivideAndRemainder(scaledDivisor, out remainder);
-			if (remainder.Sign == 0) {
-				return new BigDecimal(quotient, scale);
-			}
-			int sign = scaledDividend.Sign*scaledDivisor.Sign;
-			int compRem; // 'compare to remainder'
-			if (scaledDivisor.BitLength < 63) {
-				// 63 in order to avoid out of long after <<1
-				long rem = remainder.ToInt64();
-				long divisor = scaledDivisor.ToInt64();
-				compRem = LongCompareTo(System.Math.Abs(rem) << 1, System.Math.Abs(divisor));
-				// To look if there is a carry
-				compRem = RoundingBehavior(quotient.TestBit(0) ? 1 : 0,
-					sign*(5 + compRem), roundingMode);
-
-			} else {
-				// Checking if:  remainder * 2 >= scaledDivisor 
-				compRem = remainder.Abs().ShiftLeftOneBit().CompareTo(scaledDivisor.Abs());
-				compRem = RoundingBehavior(quotient.TestBit(0) ? 1 : 0,
-					sign*(5 + compRem), roundingMode);
-			}
-			if (compRem != 0) {
-				if (quotient.BitLength < 63) {
-					return ValueOf(quotient.ToInt64() + compRem, scale);
-				}
-				quotient = quotient.Add(BigInteger.ValueOf(compRem));
-				return new BigDecimal(quotient, scale);
-			}
-			// Constructing the result with the appropriate unscaled value
-			return new BigDecimal(quotient, scale);
-		}
-
-		private static BigDecimal DividePrimitiveLongs(long scaledDividend, long scaledDivisor, int scale,
-			RoundingMode roundingMode) {
-			long quotient = scaledDividend/scaledDivisor;
-			long remainder = scaledDividend%scaledDivisor;
-			int sign = System.Math.Sign(scaledDividend)*System.Math.Sign(scaledDivisor);
-			if (remainder != 0) {
-				// Checking if:  remainder * 2 >= scaledDivisor
-				int compRem; // 'compare to remainder'
-				compRem = LongCompareTo(System.Math.Abs(remainder) << 1, System.Math.Abs(scaledDivisor));
-				// To look if there is a carry
-				quotient += RoundingBehavior(((int) quotient) & 1, sign*(5 + compRem), roundingMode);
-			}
-			// Constructing the result with the appropriate unscaled value
-			return ValueOf(quotient, scale);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this / divisor}.
-		 * The scale of the result is the scale of {@code this}. If rounding is
-		 * required to meet the specified scale, then the specified rounding mode
-		 * {@code roundingMode} is applied.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param roundingMode
-		 *            rounding mode to be used to round the result.
-		 * @return {@code this / divisor} rounded according to the given rounding
-		 *         mode.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws IllegalArgumentException
-		 *             if {@code roundingMode} is not a valid rounding mode.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if {@code roundingMode == ROUND_UNNECESSARY} and rounding is
-		 *             necessary according to the scale of this.
-		 */
-
-		public BigDecimal Divide(BigDecimal divisor, int roundingMode) {
-			if (!Enum.IsDefined(typeof (RoundingMode), roundingMode))
-				throw new ArgumentException();
-
-			return Divide(divisor, _scale, (RoundingMode) roundingMode);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this / divisor}.
-		 * The scale of the result is the scale of {@code this}. If rounding is
-		 * required to meet the specified scale, then the specified rounding mode
-		 * {@code roundingMode} is applied.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param roundingMode
-		 *            rounding mode to be used to round the result.
-		 * @return {@code this / divisor} rounded according to the given rounding
-		 *         mode.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null} or {@code roundingMode == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if {@code roundingMode == RoundingMode.UNNECESSARY} and
-		 *             rounding is necessary according to the scale of this.
-		 */
-
-		public BigDecimal Divide(BigDecimal divisor, RoundingMode roundingMode) {
-			return Divide(divisor, _scale, roundingMode);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this / divisor}.
-		 * The scale of the result is the difference of the scales of {@code this}
-		 * and {@code divisor}. If the exact result requires more digits, then the
-		 * scale is adjusted accordingly. For example, {@code 1/128 = 0.0078125}
-		 * which has a scale of {@code 7} and precision {@code 5}.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @return {@code this / divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if the result cannot be represented exactly.
-		 */
-
-		public BigDecimal Divide(BigDecimal divisor) {
-			BigInteger p = this.GetUnscaledValue();
-			BigInteger q = divisor.GetUnscaledValue();
-			BigInteger gcd; // greatest common divisor between 'p' and 'q'
-			BigInteger quotient;
-			BigInteger remainder;
-			long diffScale = (long) _scale - divisor._scale;
-			int newScale; // the new scale for final quotient
-			int k; // number of factors "2" in 'q'
-			int l = 0; // number of factors "5" in 'q'
-			int i = 1;
-			int lastPow = FivePow.Length - 1;
-
-			if (divisor.IsZero) {
-				// math.04=Division by zero
-				throw new ArithmeticException(Messages.math04); //$NON-NLS-1$
-			}
-			if (p.Sign == 0) {
-				return GetZeroScaledBy(diffScale);
-			}
-			// To divide both by the GCD
-			gcd = p.Gcd(q);
-			p = p.Divide(gcd);
-			q = q.Divide(gcd);
-			// To simplify all "2" factors of q, dividing by 2^k
-			k = q.LowestSetBit;
-			q = q.ShiftRight(k);
-			// To simplify all "5" factors of q, dividing by 5^l
-			do {
-				quotient = q.DivideAndRemainder(FivePow[i], out remainder);
-				if (remainder.Sign == 0) {
-					l += i;
-					if (i < lastPow) {
-						i++;
-					}
-					q = quotient;
-				} else {
-					if (i == 1) {
-						break;
-					}
-					i = 1;
-				}
-			} while (true);
-			// If  abs(q) != 1  then the quotient is periodic
-			if (!q.Abs().Equals(BigInteger.One)) {
-				// math.05=Non-terminating decimal expansion; no exact representable decimal result.
-				throw new ArithmeticException(Messages.math05); //$NON-NLS-1$
-			}
-			// The sign of the is fixed and the quotient will be saved in 'p'
-			if (q.Sign < 0) {
-				p = p.Negate();
-			}
-			// Checking if the new scale is out of range
-			newScale = ToIntScale(diffScale + System.Math.Max(k, l));
-			// k >= 0  and  l >= 0  implies that  k - l  is in the 32-bit range
-			i = k - l;
-
-			p = (i > 0)
-				? Multiplication.MultiplyByFivePow(p, i)
-				: p.ShiftLeft(-i);
-			return new BigDecimal(p, newScale);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this / divisor}.
-		 * The result is rounded according to the passed context {@code mc}. If the
-		 * passed math context specifies precision {@code 0}, then this call is
-		 * equivalent to {@code this.divide(divisor)}.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param mc
-		 *            rounding mode and precision for the result of this operation.
-		 * @return {@code this / divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null} or {@code mc == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if {@code mc.getRoundingMode() == UNNECESSARY} and rounding
-		 *             is necessary according {@code mc.getPrecision()}.
-		 */
-
-		public BigDecimal Divide(BigDecimal divisor, MathContext mc) {
-			/* Calculating how many zeros must be append to 'dividend'
-			 * to obtain a  quotient with at least 'mc.precision()' digits */
-			long traillingZeros = mc.Precision + 2L
-			                      + divisor.AproxPrecision() - AproxPrecision();
-			long diffScale = (long) _scale - divisor._scale;
-			long newScale = diffScale; // scale of the final quotient
-			int compRem; // to compare the remainder
-			int i = 1; // index   
-			int lastPow = TenPow.Length - 1; // last power of ten
-			BigInteger integerQuot; // for temporal results
-			BigInteger quotient = GetUnscaledValue();
-			BigInteger remainder;
-			// In special cases it reduces the problem to call the dual method
-			if ((mc.Precision == 0) || (IsZero) || (divisor.IsZero))
-				return Divide(divisor);
-
-			if (traillingZeros > 0) {
-				// To append trailing zeros at end of dividend
-				quotient = GetUnscaledValue().Multiply(Multiplication.PowerOf10(traillingZeros));
-				newScale += traillingZeros;
-			}
-			quotient = quotient.DivideAndRemainder(divisor.GetUnscaledValue(), out remainder);
-			integerQuot = quotient;
-			// Calculating the exact quotient with at least 'mc.precision()' digits
-			if (remainder.Sign != 0) {
-				// Checking if:   2 * remainder >= divisor ?
-				compRem = remainder.ShiftLeftOneBit().CompareTo(divisor.GetUnscaledValue());
-				// quot := quot * 10 + r;     with 'r' in {-6,-5,-4, 0,+4,+5,+6}
-				integerQuot = integerQuot.Multiply(BigInteger.Ten)
-					.Add(BigInteger.ValueOf(quotient.Sign*(5 + compRem)));
-				newScale++;
-			} else {
-				// To strip trailing zeros until the preferred scale is reached
-				while (!integerQuot.TestBit(0)) {
-					quotient = integerQuot.DivideAndRemainder(TenPow[i], out remainder);
-					if ((remainder.Sign == 0)
-					    && (newScale - i >= diffScale)) {
-						newScale -= i;
-						if (i < lastPow) {
-							i++;
-						}
-						integerQuot = quotient;
-					} else {
-						if (i == 1) {
-							break;
-						}
-						i = 1;
-					}
-				}
-			}
-			// To perform rounding
-			return new BigDecimal(integerQuot, ToIntScale(newScale), mc);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the integral part of
-		 * {@code this / divisor}. The quotient is rounded down towards zero to the
-		 * next integer. For example, {@code 0.5/0.2 = 2}.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @return integral part of {@code this / divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 */
-
-		public BigDecimal DivideToIntegralValue(BigDecimal divisor) {
-			BigInteger integralValue; // the integer of result
-			BigInteger powerOfTen; // some power of ten
-			BigInteger quotient;
-			BigInteger remainder;
-			long newScale = (long) _scale - divisor._scale;
-			long tempScale = 0;
-			int i = 1;
-			int lastPow = TenPow.Length - 1;
-
-			if (divisor.IsZero) {
-				// math.04=Division by zero
-				throw new ArithmeticException(Messages.math04); //$NON-NLS-1$
-			}
-			if ((divisor.AproxPrecision() + newScale > this.AproxPrecision() + 1L)
-			    || (this.IsZero)) {
-				/* If the divisor's integer part is greater than this's integer part,
-				 * the result must be zero with the appropriate scale */
-				integralValue = BigInteger.Zero;
-			} else if (newScale == 0) {
-				integralValue = GetUnscaledValue().Divide(divisor.GetUnscaledValue());
-			} else if (newScale > 0) {
-				powerOfTen = Multiplication.PowerOf10(newScale);
-				integralValue = GetUnscaledValue().Divide(divisor.GetUnscaledValue().Multiply(powerOfTen));
-				integralValue = integralValue.Multiply(powerOfTen);
-			} else {
-// (newScale < 0)
-				powerOfTen = Multiplication.PowerOf10(-newScale);
-				integralValue = GetUnscaledValue().Multiply(powerOfTen).Divide(divisor.GetUnscaledValue());
-				// To strip trailing zeros approximating to the preferred scale
-				while (!integralValue.TestBit(0)) {
-					quotient = integralValue.DivideAndRemainder(TenPow[i], out remainder);
-					if ((remainder.Sign == 0)
-					    && (tempScale - i >= newScale)) {
-						tempScale -= i;
-						if (i < lastPow) {
-							i++;
-						}
-						integralValue = quotient;
-					} else {
-						if (i == 1) {
-							break;
-						}
-						i = 1;
-					}
-				}
-				newScale = tempScale;
-			}
-			return ((integralValue.Sign == 0)
-				? GetZeroScaledBy(newScale)
-				: new BigDecimal(integralValue, ToIntScale(newScale)));
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the integral part of
-		 * {@code this / divisor}. The quotient is rounded down towards zero to the
-		 * next integer. The rounding mode passed with the parameter {@code mc} is
-		 * not considered. But if the precision of {@code mc > 0} and the integral
-		 * part requires more digits, then an {@code ArithmeticException} is thrown.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param mc
-		 *            math context which determines the maximal precision of the
-		 *            result.
-		 * @return integral part of {@code this / divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null} or {@code mc == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if {@code mc.getPrecision() > 0} and the result requires more
-		 *             digits to be represented.
-		 */
-
-		public BigDecimal DivideToIntegralValue(BigDecimal divisor, MathContext mc) {
-			int mcPrecision = mc.Precision;
-			int diffPrecision = Precision - divisor.Precision;
-			int lastPow = TenPow.Length - 1;
-			long diffScale = (long) _scale - divisor._scale;
-			long newScale = diffScale;
-			long quotPrecision = diffPrecision - diffScale + 1;
-			BigInteger quotient;
-			BigInteger remainder;
-			// In special cases it call the dual method
-			if ((mcPrecision == 0) || (IsZero) || (divisor.IsZero)) {
-				return DivideToIntegralValue(divisor);
-			}
-			// Let be:   this = [u1,s1]   and   divisor = [u2,s2]
-			if (quotPrecision <= 0) {
-				quotient = BigInteger.Zero;
-			} else if (diffScale == 0) {
-				// CASE s1 == s2:  to calculate   u1 / u2 
-				quotient = GetUnscaledValue().Divide(divisor.GetUnscaledValue());
-			} else if (diffScale > 0) {
-				// CASE s1 >= s2:  to calculate   u1 / (u2 * 10^(s1-s2)  
-				quotient = GetUnscaledValue().Divide(divisor.GetUnscaledValue().Multiply(Multiplication.PowerOf10(diffScale)));
-				// To chose  10^newScale  to get a quotient with at least 'mc.precision()' digits
-				newScale = System.Math.Min(diffScale, System.Math.Max(mcPrecision - quotPrecision + 1, 0));
-				// To calculate: (u1 / (u2 * 10^(s1-s2)) * 10^newScale
-				quotient = quotient.Multiply(Multiplication.PowerOf10(newScale));
-			} else {
-// CASE s2 > s1:   
-				/* To calculate the minimum power of ten, such that the quotient 
-				 *   (u1 * 10^exp) / u2   has at least 'mc.precision()' digits. */
-				long exp = System.Math.Min(-diffScale, System.Math.Max((long) mcPrecision - diffPrecision, 0));
-				long compRemDiv;
-				// Let be:   (u1 * 10^exp) / u2 = [q,r]  
-				quotient = GetUnscaledValue()
-					.Multiply(Multiplication.PowerOf10(exp))
-					.DivideAndRemainder(divisor.GetUnscaledValue(), out remainder);
-				newScale += exp; // To fix the scale
-				exp = -newScale; // The remaining power of ten
-				// If after division there is a remainder...
-				if ((remainder.Sign != 0) && (exp > 0)) {
-					// Log10(r) + ((s2 - s1) - exp) > mc.precision ?
-					compRemDiv = (new BigDecimal(remainder)).Precision
-					             + exp - divisor.Precision;
-					if (compRemDiv == 0) {
-						// To calculate:  (r * 10^exp2) / u2
-						remainder = remainder.Multiply(Multiplication.PowerOf10(exp)).Divide(divisor.GetUnscaledValue());
-						compRemDiv = System.Math.Abs(remainder.Sign);
-					}
-					if (compRemDiv > 0) {
-						// The quotient won't fit in 'mc.precision()' digits
-						// math.06=Division impossible
-						throw new ArithmeticException(Messages.math06); //$NON-NLS-1$
-					}
-				}
-			}
-			// Fast return if the quotient is zero
-			if (quotient.Sign == 0) {
-				return GetZeroScaledBy(diffScale);
-			}
-			BigInteger strippedBI = quotient;
-			BigDecimal integralValue = new BigDecimal(quotient);
-			long resultPrecision = integralValue.Precision;
-			int i = 1;
-			// To strip trailing zeros until the specified precision is reached
-			while (!strippedBI.TestBit(0)) {
-				quotient = strippedBI.DivideAndRemainder(TenPow[i], out remainder);
-				if ((remainder.Sign == 0) &&
-				    ((resultPrecision - i >= mcPrecision)
-				     || (newScale - i >= diffScale))) {
-					resultPrecision -= i;
-					newScale -= i;
-					if (i < lastPow) {
-						i++;
-					}
-					strippedBI = quotient;
-				} else {
-					if (i == 1) {
-						break;
-					}
-					i = 1;
-				}
-			}
-			// To check if the result fit in 'mc.precision()' digits
-			if (resultPrecision > mcPrecision) {
-				// math.06=Division impossible
-				throw new ArithmeticException(Messages.math06); //$NON-NLS-1$
-			}
-			integralValue._scale = ToIntScale(newScale);
-			integralValue.SetUnscaledValue(strippedBI);
-			return integralValue;
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this % divisor}.
-		 * <p>
-		 * The remainder is defined as {@code this -
-		 * this.divideToIntegralValue(divisor) * divisor}.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @return {@code this % divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 */
-
-		public BigDecimal Remainder(BigDecimal divisor) {
-			BigDecimal remainder;
-			DivideAndRemainder(divisor, out remainder);
-			return remainder;
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this % divisor}.
-		 * <p>
-		 * The remainder is defined as {@code this -
-		 * this.divideToIntegralValue(divisor) * divisor}.
-		 * <p>
-		 * The specified rounding mode {@code mc} is used for the division only.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param mc
-		 *            rounding mode and precision to be used.
-		 * @return {@code this % divisor}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @throws ArithmeticException
-		 *             if {@code mc.getPrecision() > 0} and the result of {@code
-		 *             this.divideToIntegralValue(divisor, mc)} requires more digits
-		 *             to be represented.
-		 */
-
-		public BigDecimal Remainder(BigDecimal divisor, MathContext mc) {
-			BigDecimal remainder;
-			DivideAndRemainder(divisor, mc, out remainder);
-			return remainder;
-		}
-
-		/**
-		 * Returns a {@code BigDecimal} array which contains the integral part of
-		 * {@code this / divisor} at index 0 and the remainder {@code this %
-		 * divisor} at index 1. The quotient is rounded down towards zero to the
-		 * next integer.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @return {@code [this.divideToIntegralValue(divisor),
-		 *         this.remainder(divisor)]}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @see #divideToIntegralValue
-		 * @see #remainder
-		 */
-
-		public BigDecimal DivideAndRemainder(BigDecimal divisor, out BigDecimal remainder) {
-			var quotient = DivideToIntegralValue(divisor);
-			remainder = Subtract(quotient.Multiply(divisor));
-			return quotient;
-		}
-
-		/**
-		 * Returns a {@code BigDecimal} array which contains the integral part of
-		 * {@code this / divisor} at index 0 and the remainder {@code this %
-		 * divisor} at index 1. The quotient is rounded down towards zero to the
-		 * next integer. The rounding mode passed with the parameter {@code mc} is
-		 * not considered. But if the precision of {@code mc > 0} and the integral
-		 * part requires more digits, then an {@code ArithmeticException} is thrown.
-		 *
-		 * @param divisor
-		 *            value by which {@code this} is divided.
-		 * @param mc
-		 *            math context which determines the maximal precision of the
-		 *            result.
-		 * @return {@code [this.divideToIntegralValue(divisor),
-		 *         this.remainder(divisor)]}.
-		 * @throws NullPointerException
-		 *             if {@code divisor == null}.
-		 * @throws ArithmeticException
-		 *             if {@code divisor == 0}.
-		 * @see #divideToIntegralValue
-		 * @see #remainder
-		 */
-
-		public BigDecimal DivideAndRemainder(BigDecimal divisor, MathContext mc, out BigDecimal remainder) {
-			var quotient = DivideToIntegralValue(divisor, mc);
-			remainder = Subtract(quotient.Multiply(divisor));
-			return quotient;
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this ^ n}. The
-		 * scale of the result is {@code n} times the scales of {@code this}.
-		 * <p>
-		 * {@code x.pow(0)} returns {@code 1}, even if {@code x == 0}.
-		 * <p>
-		 * Implementation Note: The implementation is based on the ANSI standard
-		 * X3.274-1996 algorithm.
-		 *
-		 * @param n
-		 *            exponent to which {@code this} is raised.
-		 * @return {@code this ^ n}.
-		 * @throws ArithmeticException
-		 *             if {@code n < 0} or {@code n > 999999999}.
-		 */
-
-		public BigDecimal Pow(int n) {
-			if (n == 0) {
-				return One;
-			}
-			if ((n < 0) || (n > 999999999)) {
-				// math.07=Invalid Operation
-				throw new ArithmeticException(Messages.math07); //$NON-NLS-1$
-			}
-			long newScale = _scale*(long) n;
-			// Let be: this = [u,s]   so:  this^n = [u^n, s*n]
-			return ((IsZero)
-				? GetZeroScaledBy(newScale)
-				: new BigDecimal(GetUnscaledValue().Pow(n), ToIntScale(newScale)));
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this ^ n}. The
-		 * result is rounded according to the passed context {@code mc}.
-		 * <p>
-		 * Implementation Note: The implementation is based on the ANSI standard
-		 * X3.274-1996 algorithm.
-		 *
-		 * @param n
-		 *            exponent to which {@code this} is raised.
-		 * @param mc
-		 *            rounding mode and precision for the result of this operation.
-		 * @return {@code this ^ n}.
-		 * @throws ArithmeticException
-		 *             if {@code n < 0} or {@code n > 999999999}.
-		 */
-
-		public BigDecimal Pow(int n, MathContext mc) {
-			// The ANSI standard X3.274-1996 algorithm
-			int m = System.Math.Abs(n);
-			int mcPrecision = mc.Precision;
-			int elength = (int) System.Math.Log10(m) + 1; // decimal digits in 'n'
-			int oneBitMask; // mask of bits
-			BigDecimal accum; // the single accumulator
-			MathContext newPrecision = mc; // MathContext by default
-
-			// In particular cases, it reduces the problem to call the other 'pow()'
-			if ((n == 0) || ((IsZero) && (n > 0))) {
-				return Pow(n);
-			}
-			if ((m > 999999999) || ((mcPrecision == 0) && (n < 0))
-			    || ((mcPrecision > 0) && (elength > mcPrecision))) {
-				// math.07=Invalid Operation
-				throw new ArithmeticException(Messages.math07); //$NON-NLS-1$
-			}
-			if (mcPrecision > 0) {
-				newPrecision = new MathContext(mcPrecision + elength + 1,
-					mc.RoundingMode);
-			}
-			// The result is calculated as if 'n' were positive        
-			accum = Round(newPrecision);
-			oneBitMask = Utils.HighestOneBit(m) >> 1;
-
-			while (oneBitMask > 0) {
-				accum = accum.Multiply(accum, newPrecision);
-				if ((m & oneBitMask) == oneBitMask) {
-					accum = accum.Multiply(this, newPrecision);
-				}
-				oneBitMask >>= 1;
-			}
-			// If 'n' is negative, the value is divided into 'ONE'
-			if (n < 0) {
-				accum = One.Divide(accum, newPrecision);
-			}
-			// The final value is rounded to the destination precision
-			accum.InplaceRound(mc);
-			return accum;
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the absolute value of
-		 * {@code this}. The scale of the result is the same as the scale of this.
-		 *
-		 * @return {@code abs(this)}
-		 */
-
-		public BigDecimal Abs() {
-			return ((Sign < 0) ? Negate() : this);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the absolute value of
-		 * {@code this}. The result is rounded according to the passed context
-		 * {@code mc}.
-		 *
-		 * @param mc
-		 *            rounding mode and precision for the result of this operation.
-		 * @return {@code abs(this)}
-		 */
-
-		public BigDecimal Abs(MathContext mc) {
-			return Round(mc).Abs();
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the {@code -this}. The
-		 * scale of the result is the same as the scale of this.
-		 *
-		 * @return {@code -this}
-		 */
-
-		public BigDecimal Negate() {
-			if (_bitLength < 63 || (_bitLength == 63 && smallValue != Int64.MinValue)) {
-				return ValueOf(-smallValue, _scale);
-			}
-			return new BigDecimal(GetUnscaledValue().Negate(), _scale);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is the {@code -this}. The
-		 * result is rounded according to the passed context {@code mc}.
-		 *
-		 * @param mc
-		 *            rounding mode and precision for the result of this operation.
-		 * @return {@code -this}
-		 */
-
-		public BigDecimal Negate(MathContext mc) {
-			return Round(mc).Negate();
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code +this}. The scale
-		 * of the result is the same as the scale of this.
-		 *
-		 * @return {@code this}
-		 */
-
-		public BigDecimal Plus() {
-			return this;
-		}
-
-		/// <remarks>
-		/// Returns a new <see cref="BigDecimal"/> whose value is <c>+this</c>.
-		/// </remarks>
-		/// <param name="mc">Rounding mode and precision for the result of this operation.</param>
-		/// <remarks>
-		/// The result is rounded according to the passed context <paramref name="mc"/>.
-		/// </remarks>
-		/// <returns>
-		/// Returns this decimal value rounded.
-		/// </returns>
-		public BigDecimal Plus(MathContext mc) {
-			return Round(mc);
-		}
-
-		public BigDecimal Round(MathContext mc) {
-			var thisBD = new BigDecimal(GetUnscaledValue(), _scale);
-
-			thisBD.InplaceRound(mc);
-			return thisBD;
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance with the specified scale.
-		 * <p>
-		 * If the new scale is greater than the old scale, then additional zeros are
-		 * added to the unscaled value. In this case no rounding is necessary.
-		 * <p>
-		 * If the new scale is smaller than the old scale, then trailing digits are
-		 * removed. If these trailing digits are not zero, then the remaining
-		 * unscaled value has to be rounded. For this rounding operation the
-		 * specified rounding mode is used.
-		 *
-		 * @param newScale
-		 *            scale of the result returned.
-		 * @param roundingMode
-		 *            rounding mode to be used to round the result.
-		 * @return a new {@code BigDecimal} instance with the specified scale.
-		 * @throws NullPointerException
-		 *             if {@code roundingMode == null}.
-		 * @throws ArithmeticException
-		 *             if {@code roundingMode == ROUND_UNNECESSARY} and rounding is
-		 *             necessary according to the given scale.
-		 */
-
-		public BigDecimal SetScale(int newScale, RoundingMode roundingMode) {
-			long diffScale = newScale - (long) _scale;
-			// Let be:  'this' = [u,s]        
-			if (diffScale == 0) {
-				return this;
-			}
-			if (diffScale > 0) {
-				// return  [u * 10^(s2 - s), newScale]
-				if (diffScale < LongTenPow.Length &&
-				    (_bitLength + LongTenPowBitLength[(int) diffScale]) < 64) {
-					return ValueOf(smallValue*LongTenPow[(int) diffScale], newScale);
-				}
-				return new BigDecimal(Multiplication.MultiplyByTenPow(GetUnscaledValue(), (int) diffScale), newScale);
-			}
-			// diffScale < 0
-			// return  [u,s] / [1,newScale]  with the appropriate scale and rounding
-			if (_bitLength < 64 && -diffScale < LongTenPow.Length) {
-				return DividePrimitiveLongs(smallValue, LongTenPow[(int) -diffScale], newScale, roundingMode);
-			}
-			return DivideBigIntegers(GetUnscaledValue(), Multiplication.PowerOf10(-diffScale), newScale, roundingMode);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance with the specified scale.
-		 * <p>
-		 * If the new scale is greater than the old scale, then additional zeros are
-		 * added to the unscaled value. In this case no rounding is necessary.
-		 * <p>
-		 * If the new scale is smaller than the old scale, then trailing digits are
-		 * removed. If these trailing digits are not zero, then the remaining
-		 * unscaled value has to be rounded. For this rounding operation the
-		 * specified rounding mode is used.
-		 *
-		 * @param newScale
-		 *            scale of the result returned.
-		 * @param roundingMode
-		 *            rounding mode to be used to round the result.
-		 * @return a new {@code BigDecimal} instance with the specified scale.
-		 * @throws IllegalArgumentException
-		 *             if {@code roundingMode} is not a valid rounding mode.
-		 * @throws ArithmeticException
-		 *             if {@code roundingMode == ROUND_UNNECESSARY} and rounding is
-		 *             necessary according to the given scale.
-		 */
-
-		public BigDecimal SetScale(int newScale, int roundingMode) {
-			RoundingMode rm = (RoundingMode) roundingMode;
-			if ((roundingMode < (int) RoundingMode.Up) ||
-			    (roundingMode > (int) RoundingMode.Unnecessary)) {
-				throw new ArgumentException("roundingMode");
-			}
-			return SetScale(newScale, (RoundingMode) roundingMode);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance with the specified scale. If
-		 * the new scale is greater than the old scale, then additional zeros are
-		 * added to the unscaled value. If the new scale is smaller than the old
-		 * scale, then trailing zeros are removed. If the trailing digits are not
-		 * zeros then an ArithmeticException is thrown.
-		 * <p>
-		 * If no exception is thrown, then the following equation holds: {@code
-		 * x.setScale(s).compareTo(x) == 0}.
-		 *
-		 * @param newScale
-		 *            scale of the result returned.
-		 * @return a new {@code BigDecimal} instance with the specified scale.
-		 * @throws ArithmeticException
-		 *             if rounding would be necessary.
-		 */
-
-		public BigDecimal SetScale(int newScale) {
-			return SetScale(newScale, RoundingMode.Unnecessary);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance where the decimal point has
-		 * been moved {@code n} places to the left. If {@code n < 0} then the
-		 * decimal point is moved {@code -n} places to the right.
-		 * <p>
-		 * The result is obtained by changing its scale. If the scale of the result
-		 * becomes negative, then its precision is increased such that the scale is
-		 * zero.
-		 * <p>
-		 * Note, that {@code movePointLeft(0)} returns a result which is
-		 * mathematically equivalent, but which has {@code scale >= 0}.
-		 *
-		 * @param n
-		 *            number of placed the decimal point has to be moved.
-		 * @return {@code this * 10^(-n}).
-		 */
-
-		public BigDecimal MovePointLeft(int n) {
-			return MovePoint(_scale + (long) n);
-		}
-
-		private BigDecimal MovePoint(long newScale) {
-			if (IsZero) {
-				return GetZeroScaledBy(System.Math.Max(newScale, 0));
-			}
-			/* When:  'n'== Integer.MIN_VALUE  isn't possible to call to movePointRight(-n)  
-			 * since  -Integer.MIN_VALUE == Integer.MIN_VALUE */
-			if (newScale >= 0) {
-				if (_bitLength < 64) {
-					return ValueOf(smallValue, ToIntScale(newScale));
-				}
-				return new BigDecimal(GetUnscaledValue(), ToIntScale(newScale));
-			}
-			if (-newScale < LongTenPow.Length &&
-			    _bitLength + LongTenPowBitLength[(int) -newScale] < 64) {
-				return ValueOf(smallValue*LongTenPow[(int) -newScale], 0);
-			}
-			return new BigDecimal(Multiplication.MultiplyByTenPow(GetUnscaledValue(), (int) -newScale), 0);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance where the decimal point has
-		 * been moved {@code n} places to the right. If {@code n < 0} then the
-		 * decimal point is moved {@code -n} places to the left.
-		 * <p>
-		 * The result is obtained by changing its scale. If the scale of the result
-		 * becomes negative, then its precision is increased such that the scale is
-		 * zero.
-		 * <p>
-		 * Note, that {@code movePointRight(0)} returns a result which is
-		 * mathematically equivalent, but which has scale >= 0.
-		 *
-		 * @param n
-		 *            number of placed the decimal point has to be moved.
-		 * @return {@code this * 10^n}.
-		 */
-
-		public BigDecimal MovePointRight(int n) {
-			return MovePoint(_scale - (long) n);
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} whose value is {@code this} 10^{@code n}.
-		 * The scale of the result is {@code this.scale()} - {@code n}.
-		 * The precision of the result is the precision of {@code this}.
-		 * <p>
-		 * This method has the same effect as {@link #movePointRight}, except that
-		 * the precision is not changed.
-		 *
-		 * @param n
-		 *            number of places the decimal point has to be moved.
-		 * @return {@code this * 10^n}
-		 */
-
-		public BigDecimal ScaleByPowerOfTen(int n) {
-			long newScale = _scale - (long) n;
-			if (_bitLength < 64) {
-				//Taking care when a 0 is to be scaled
-				if (smallValue == 0) {
-					return GetZeroScaledBy(newScale);
-				}
-				return ValueOf(smallValue, ToIntScale(newScale));
-			}
-			return new BigDecimal(GetUnscaledValue(), ToIntScale(newScale));
-		}
-
-		/**
-		 * Returns a new {@code BigDecimal} instance with the same value as {@code
-		 * this} but with a unscaled value where the trailing zeros have been
-		 * removed. If the unscaled value of {@code this} has n trailing zeros, then
-		 * the scale and the precision of the result has been reduced by n.
-		 *
-		 * @return a new {@code BigDecimal} instance equivalent to this where the
-		 *         trailing zeros of the unscaled value have been removed.
-		 */
-
-		public BigDecimal StripTrailingZeros() {
-			int i = 1; // 1 <= i <= 18
-			int lastPow = TenPow.Length - 1;
-			long newScale = _scale;
-
-			if (IsZero) {
-				return BigDecimal.Parse("0");
-			}
-			BigInteger strippedBI = GetUnscaledValue();
-			BigInteger quotient;
-			BigInteger remainder;
-
-			// while the number is even...
-			while (!strippedBI.TestBit(0)) {
-				// To divide by 10^i
-				quotient = strippedBI.DivideAndRemainder(TenPow[i], out remainder);
-				// To look the remainder
-				if (remainder.Sign == 0) {
-					// To adjust the scale
-					newScale -= i;
-					if (i < lastPow) {
-						// To set to the next power
-						i++;
-					}
-					strippedBI = quotient;
-				} else {
-					if (i == 1) {
-						// 'this' has no more trailing zeros
-						break;
-					}
-					// To set to the smallest power of ten
-					i = 1;
-				}
-			}
-			return new BigDecimal(strippedBI, ToIntScale(newScale));
-		}
-
-		public BigDecimal Min(BigDecimal val) {
-			return ((CompareTo(val) <= 0) ? this : val);
-		}
-
-		/**
-		 * Returns the maximum of this {@code BigDecimal} and {@code val}.
-		 *
-		 * @param val
-		 *            value to be used to compute the maximum with this.
-		 * @return {@code max(this, val}.
-		 * @throws NullPointerException
-		 *             if {@code val == null}.
-		 */
-
-		public BigDecimal Max(BigDecimal val) {
-			return ((CompareTo(val) >= 0) ? this : val);
-		}
 
 		#endregion
 
@@ -2118,7 +1150,7 @@ namespace Deveel.Math {
 				compRem = RoundingBehavior(integer.TestBit(0) ? 1 : 0, fraction.Sign*(5 + compRem),
 					mc.RoundingMode);
 				if (compRem != 0) {
-					integer = integer.Add(BigInteger.ValueOf(compRem));
+					integer = integer.Add(BigInteger.FromInt64(compRem));
 				}
 				tempBD = new BigDecimal(integer);
 				// If after to add the increment the precision changed, we normalize the size
@@ -2323,7 +1355,7 @@ namespace Deveel.Math {
 
 		private BigInteger GetUnscaledValue() {
 			if (intVal == null)
-				intVal = BigInteger.ValueOf(smallValue);
+				intVal = BigInteger.FromInt64(smallValue);
 			return intVal;
 		}
 
@@ -2382,1013 +1414,5 @@ namespace Deveel.Math {
 		}
 
 #endif
-
-		#region Conversions
-
-#if !PORTABLE
-		TypeCode IConvertible.GetTypeCode() {
-			return TypeCode.Object;
-		}
-
-		bool IConvertible.ToBoolean(IFormatProvider provider) {
-			int value = ToInt32();
-			if (value == 1)
-				return true;
-			if (value == 0)
-				return false;
-			throw new InvalidCastException();
-		}
-
-		char IConvertible.ToChar(IFormatProvider provider) {
-			short value = ToInt16Exact();
-			return (char) value;
-		}
-
-		sbyte IConvertible.ToSByte(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		byte IConvertible.ToByte(IFormatProvider provider) {
-			int value = ToInt32();
-			if (value > Byte.MaxValue || value < Byte.MinValue)
-				throw new InvalidCastException();
-			return (byte) value;
-		}
-
-		short IConvertible.ToInt16(IFormatProvider provider) {
-			return ToInt16Exact();
-		}
-
-		ushort IConvertible.ToUInt16(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		int IConvertible.ToInt32(IFormatProvider provider) {
-			return ToInt32();
-		}
-
-		uint IConvertible.ToUInt32(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		long IConvertible.ToInt64(IFormatProvider provider) {
-			return ToInt64();
-		}
-
-		ulong IConvertible.ToUInt64(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		float IConvertible.ToSingle(IFormatProvider provider) {
-			return ToSingle();
-		}
-
-		double IConvertible.ToDouble(IFormatProvider provider) {
-			return ToDouble();
-		}
-
-		decimal IConvertible.ToDecimal(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		DateTime IConvertible.ToDateTime(IFormatProvider provider) {
-			throw new NotSupportedException();
-		}
-
-		string IConvertible.ToString(IFormatProvider provider) {
-			return ToString(provider);
-		}
-
-		object IConvertible.ToType(Type conversionType, IFormatProvider provider) {
-			if (conversionType == typeof(bool))
-				return (this as IConvertible).ToBoolean(provider);
-			if (conversionType == typeof(byte))
-				return (this as IConvertible).ToByte(provider);
-			if (conversionType == typeof(short))
-				return ToInt16Exact();
-			if (conversionType == typeof(int))
-				return ToInt32();
-			if (conversionType == typeof(long))
-				return ToInt64();
-			if (conversionType == typeof(float))
-				return ToSingle();
-			if (conversionType == typeof(double))
-				return ToDouble();
-
-			if (conversionType == typeof(BigInteger))
-				return ToBigInteger();
-
-			throw new NotSupportedException();
-		}
-
-		#endif
-
-		public override string ToString() {
-			if (toStringImage != null) {
-				return toStringImage;
-			}
-			return ToString(null);
-		}
-
-		public string ToString(IFormatProvider provider) {
-			if (provider == null)
-				provider = NumberFormatInfo.InvariantInfo;
-
-			return ToStringInternal(provider);
-		}
-
-		private string ToStringInternal(IFormatProvider provider) {
-			var numberInfo = provider.GetFormat(typeof (NumberFormatInfo)) as NumberFormatInfo;
-			if (numberInfo == null)
-				numberInfo = NumberFormatInfo.InvariantInfo;
-
-			var decimalSep = numberInfo.NumberDecimalSeparator;
-			if (decimalSep.Length > 1)
-				throw new NotSupportedException("Decimal separators with more than one character are not supported (yet).");
-
-			if (_bitLength < 32) {
-				toStringImage = Conversion.ToDecimalScaledString(smallValue, _scale);
-				return toStringImage;
-			}
-			String intString = GetUnscaledValue().ToString();
-			if (_scale == 0) {
-				return intString;
-			}
-			int begin = (GetUnscaledValue().Sign < 0) ? 2 : 1;
-			int end = intString.Length;
-			long exponent = -(long) _scale + end - begin;
-			StringBuilder result = new StringBuilder();
-
-			result.Append(intString);
-			if ((_scale > 0) && (exponent >= -6)) {
-				if (exponent >= 0) {
-					result.Insert(end - _scale, decimalSep);
-				} else {
-					result.Insert(begin - 1, "0" + decimalSep); //$NON-NLS-1$
-					result.Insert(begin + 1, ChZeros, 0, -(int) exponent - 1);
-				}
-			} else {
-				if (end - begin >= 1) {
-					result.Insert(begin, decimalSep);
-					end++;
-				}
-				result.Insert(end, new[] {'E'});
-				if (exponent > 0) {
-					result.Insert(++end, new[] {'+'});
-				}
-				result.Insert(++end, Convert.ToString(exponent));
-			}
-			toStringImage = result.ToString();
-			return toStringImage;
-		}
-
-		/**
-		 * Returns a string representation of this {@code BigDecimal}. This
-		 * representation always prints all significant digits of this value.
-		 * <p>
-		 * If the scale is negative or if {@code scale - precision >= 6} then
-		 * engineering notation is used. Engineering notation is similar to the
-		 * scientific notation except that the exponent is made to be a multiple of
-		 * 3 such that the integer part is >= 1 and < 1000.
-		 *
-		 * @return a string representation of {@code this} in engineering notation
-		 *         if necessary.
-		 */
-
-		public String ToEngineeringString() {
-			return ToEngineeringString(null);
-		}
-
-		public String ToEngineeringString(IFormatProvider provider) {
-			var numberInfo = provider.GetFormat(typeof (NumberFormatInfo)) as NumberFormatInfo;
-			if (numberInfo == null)
-				numberInfo = NumberFormatInfo.InvariantInfo;
-
-			var decimalSep = numberInfo.NumberDecimalSeparator;
-			if (decimalSep.Length > 1)
-				throw new NotSupportedException("Decimal separators with more than one character are not supported (yet).");
-
-			String intString = GetUnscaledValue().ToString();
-			if (_scale == 0) {
-				return intString;
-			}
-			int begin = (GetUnscaledValue().Sign < 0) ? 2 : 1;
-			int end = intString.Length;
-			long exponent = -(long) _scale + end - begin;
-			StringBuilder result = new StringBuilder(intString);
-
-			if ((_scale > 0) && (exponent >= -6)) {
-				if (exponent >= 0) {
-					result.Insert(end - _scale, decimalSep);
-				} else {
-					result.Insert(begin - 1, "0" + decimalSep); //$NON-NLS-1$
-					result.Insert(begin + 1, ChZeros, 0, -(int) exponent - 1);
-				}
-			} else {
-				int delta = end - begin;
-				int rem = (int) (exponent%3);
-
-				if (rem != 0) {
-					// adjust exponent so it is a multiple of three
-					if (GetUnscaledValue().Sign == 0) {
-						// zero value
-						rem = (rem < 0) ? -rem : 3 - rem;
-						exponent += rem;
-					} else {
-						// nonzero value
-						rem = (rem < 0) ? rem + 3 : rem;
-						exponent -= rem;
-						begin += rem;
-					}
-					if (delta < 3) {
-						for (int i = rem - delta; i > 0; i--) {
-							result.Insert(end++, new[] {'0'});
-						}
-					}
-				}
-				if (end - begin >= 1) {
-					result.Insert(begin, decimalSep);
-					end++;
-				}
-				if (exponent != 0) {
-					result.Insert(end, new[] {'E'});
-					if (exponent > 0) {
-						result.Insert(++end, new[] {'+'});
-					}
-					result.Insert(++end, Convert.ToString(exponent));
-				}
-			}
-			return result.ToString();
-		}
-
-		/**
-		 * Returns a string representation of this {@code BigDecimal}. No scientific
-		 * notation is used. This methods adds zeros where necessary.
-		 * <p>
-		 * If this string representation is used to create a new instance, this
-		 * instance is generally not identical to {@code this} as the precision
-		 * changes.
-		 * <p>
-		 * {@code x.equals(new BigDecimal(x.toPlainString())} usually returns
-		 * {@code false}.
-		 * <p>
-		 * {@code x.compareTo(new BigDecimal(x.toPlainString())} returns {@code 0}.
-		 *
-		 * @return a string representation of {@code this} without exponent part.
-		 */
-
-		public String ToPlainString() {
-			String intStr = GetUnscaledValue().ToString();
-			if ((_scale == 0) || ((IsZero) && (_scale < 0))) {
-				return intStr;
-			}
-			int begin = (Sign < 0) ? 1 : 0;
-			int delta = _scale;
-			// We take space for all digits, plus a possible decimal point, plus 'scale'
-			StringBuilder result = new StringBuilder(intStr.Length + 1 + System.Math.Abs(_scale));
-
-			if (begin == 1) {
-				// If the number is negative, we insert a '-' CharHelper at front 
-				result.Append('-');
-			}
-			if (_scale > 0) {
-				delta -= (intStr.Length - begin);
-				if (delta >= 0) {
-					result.Append("0."); //$NON-NLS-1$
-					// To append zeros after the decimal point
-					for (; delta > ChZeros.Length; delta -= ChZeros.Length) {
-						result.Append(ChZeros);
-					}
-					result.Append(ChZeros, 0, delta);
-					result.Append(intStr.Substring(begin));
-				} else {
-					delta = begin - delta;
-					result.Append(intStr.Substring(begin, delta - begin));
-					result.Append('.');
-					result.Append(intStr.Substring(delta));
-				}
-			} else {
-// (scale <= 0)
-				result.Append(intStr.Substring(begin));
-				// To append trailing zeros
-				for (; delta < -ChZeros.Length; delta += ChZeros.Length) {
-					result.Append(ChZeros);
-				}
-				result.Append(ChZeros, 0, -delta);
-			}
-			return result.ToString();
-		}
-
-		private static bool TryParse(char[] inData, int offset, int len, IFormatProvider provider, out BigDecimal value,
-			out Exception exception) {
-			if (inData == null || inData.Length == 0) {
-				exception = new FormatException("Cannot parse an empty string.");
-				value = null;
-				return false;
-			}
-
-			var numberformatInfo = provider.GetFormat(typeof (NumberFormatInfo)) as NumberFormatInfo;
-			if (numberformatInfo == null)
-				numberformatInfo = NumberFormatInfo.CurrentInfo;
-
-			var decSep = numberformatInfo.NumberDecimalSeparator;
-			if (decSep.Length > 1) {
-				exception = new NotSupportedException("More than one decimal separator not yet supported");
-				value = null;
-				return false;
-			}
-
-			var cDecSep = decSep[0];
-
-			int begin = offset; // first index to be copied
-			int last = offset + (len - 1); // last index to be copied
-
-			if ((last >= inData.Length) || (offset < 0) || (len <= 0) || (last < 0)) {
-				exception = new FormatException();
-				value = null;
-				return false;
-			}
-
-			var v = new BigDecimal();
-
-			try {
-				var unscaledBuffer = new StringBuilder(len);
-				int bufLength = 0;
-				// To skip a possible '+' symbol
-				if ((offset <= last) && (inData[offset] == '+')) {
-					offset++;
-					begin++;
-				}
-
-				int counter = 0;
-				bool wasNonZero = false;
-				// Accumulating all digits until a possible decimal point
-				for (;
-					(offset <= last) &&
-					(inData[offset] != cDecSep) &&
-					(inData[offset] != 'e') &&
-					(inData[offset] != 'E');
-					offset++) {
-					if (!wasNonZero) {
-						if (inData[offset] == '0') {
-							counter++;
-						} else {
-							wasNonZero = true;
-						}
-					}
-				}
-
-				unscaledBuffer.Append(inData, begin, offset - begin);
-				bufLength += offset - begin;
-				// A decimal point was found
-				if ((offset <= last) && (inData[offset] == cDecSep)) {
-					offset++;
-					// Accumulating all digits until a possible exponent
-					begin = offset;
-					for (;
-						(offset <= last) &&
-						(inData[offset] != 'e') &&
-						(inData[offset] != 'E');
-						offset++) {
-						if (!wasNonZero) {
-							if (inData[offset] == '0') {
-								counter++;
-							} else {
-								wasNonZero = true;
-							}
-						}
-					}
-
-					v._scale = offset - begin;
-					bufLength += v._scale;
-					unscaledBuffer.Append(inData, begin, v._scale);
-				} else {
-					v._scale = 0;
-				}
-				// An exponent was found
-				if ((offset <= last) && ((inData[offset] == 'e') || (inData[offset] == 'E'))) {
-					offset++;
-					// Checking for a possible sign of scale
-					begin = offset;
-					if ((offset <= last) && (inData[offset] == '+')) {
-						offset++;
-						if ((offset <= last) && (inData[offset] != '-')) {
-							begin++;
-						}
-					}
-
-					// Accumulating all remaining digits
-					String scaleString = new String(inData, begin, last + 1 - begin); // buffer for scale
-					// Checking if the scale is defined            
-					long newScale = (long) v._scale - Int32.Parse(scaleString, provider); // the new scale
-					v._scale = (int) newScale;
-					if (newScale != v._scale) {
-						// math.02=Scale out of range.
-						throw new FormatException(Messages.math02); //$NON-NLS-1$
-					}
-				}
-
-				// Parsing the unscaled value
-				if (bufLength < 19) {
-					if (!Int64.TryParse(unscaledBuffer.ToString(), NumberStyles.Integer, provider, out v.smallValue)) {
-						value = null;
-						exception = new FormatException();
-						return false;
-					}
-
-					v._bitLength = BitLength(v.smallValue);
-				} else {
-					v.SetUnscaledValue(BigInteger.Parse(unscaledBuffer.ToString()));
-				}
-
-				v._precision = unscaledBuffer.Length - counter;
-				if (unscaledBuffer[0] == '-') {
-					v._precision--;
-				}
-
-				value = v;
-				exception = null;
-				return true;
-			} catch (Exception ex) {
-				exception = ex;
-				value = null;
-				return false;
-			}
-		}
-
-		public static bool TryParse(char[] chars, int offset, int length, out BigDecimal value) {
-			return TryParse(chars, offset, length, NumberFormatInfo.InvariantInfo, out value);
-		}
-
-		public static bool TryParse(char[] chars, int offset, int length, IFormatProvider provider, out BigDecimal value) {
-			return TryParse(chars, offset, length, null, provider, out value);
-		}
-
-		public static bool TryParse(char[] chars, int offset, int length, MathContext context, out BigDecimal value) {
-			return TryParse(chars, offset, length, context, null, out value);
-		}
-
-		public static bool TryParse(char[] chars, int offset, int length, MathContext context, IFormatProvider provider,
-			out BigDecimal value) {
-			Exception error;
-			if (!TryParse(chars, offset, length, provider, out value, out error))
-				return false;
-
-			if (context != null)
-				value.InplaceRound(context);
-
-			return true;
-		}
-
-		public static bool TryParse(char[] chars, out BigDecimal value) {
-			return TryParse(chars, (MathContext) null, out value);
-		}
-
-		public static bool TryParse(char[] chars, MathContext context, out BigDecimal value) {
-			return TryParse(chars, context, NumberFormatInfo.InvariantInfo, out value);
-		}
-
-		public static bool TryParse(char[] chars, IFormatProvider provider, out BigDecimal value) {
-			return TryParse(chars, null, provider, out value);
-		}
-
-		public static bool TryParse(char[] chars, MathContext context, IFormatProvider provider, out BigDecimal value) {
-			if (chars == null) {
-				value = null;
-				return false;
-			}
-
-			return TryParse(chars, 0, chars.Length, context, provider, out value);
-		}
-
-		public static BigDecimal Parse(char[] chars, int offset, int length, IFormatProvider provider) {
-			return Parse(chars, offset, length, null, provider);
-		}
-
-		public static BigDecimal Parse(char[] chars, int offset, int length) {
-			return Parse(chars, offset, length, (MathContext) null);
-		}
-
-		public static BigDecimal Parse(char[] chars, int offset, int length, MathContext context) {
-			return Parse(chars, offset, length, context, NumberFormatInfo.InvariantInfo);
-		}
-
-		public static BigDecimal Parse(char[] chars, int offset, int length, MathContext context, IFormatProvider provider) {
-			Exception error;
-			BigDecimal value;
-			if (!TryParse(chars, offset, length, provider, out value, out error))
-				throw error;
-
-			if (context != null)
-				value.InplaceRound(context);
-
-			return value;
-		}
-
-		public static BigDecimal Parse(char[] chars, IFormatProvider provider) {
-			return Parse(chars, null, provider);
-		}
-
-		public static BigDecimal Parse(char[] chars) {
-			return Parse(chars, (MathContext) null);
-		}
-
-		public static BigDecimal Parse(char[] chars, MathContext context) {
-			return Parse(chars, context, NumberFormatInfo.InvariantInfo);
-		}
-
-		public static BigDecimal Parse(char[] chars, MathContext context, IFormatProvider provider) {
-			if (chars == null)
-				throw new ArgumentNullException("chars");
-
-			return Parse(chars, 0, chars.Length, context, provider);
-		}
-
-		public static bool TryParse(string s, out BigDecimal value) {
-			return TryParse(s, (MathContext) null, out value);
-		}
-
-		public static bool TryParse(string s, MathContext context, out BigDecimal value) {
-			return TryParse(s, context, NumberFormatInfo.InvariantInfo, out value);
-		}
-
-		public static bool TryParse(string s, IFormatProvider provider, out BigDecimal value) {
-			return TryParse(s, null, provider, out value);
-		}
-
-		public static bool TryParse(string s, MathContext context, IFormatProvider provider, out BigDecimal value) {
-			if (String.IsNullOrEmpty(s)) {
-				value = null;
-				return false;
-			}
-
-			var data = s.ToCharArray();
-
-			Exception error;
-			if (!TryParse(data, 0, data.Length, provider, out value, out error))
-				return false;
-
-			if (context != null)
-				value.InplaceRound(context);
-
-			return true;
-		}
-
-		public static BigDecimal Parse(string s, IFormatProvider provider) {
-			return Parse(s, null, provider);
-		}
-
-		public static BigDecimal Parse(string s) {
-			return Parse(s, (MathContext) null);
-		}
-
-		public static BigDecimal Parse(string s, MathContext context) {
-			return Parse(s, context, NumberFormatInfo.InvariantInfo);
-		}
-
-		public static BigDecimal Parse(string s, MathContext context, IFormatProvider provider) {
-			if (String.IsNullOrEmpty(s))
-				throw new FormatException();
-
-			var data = s.ToCharArray();
-
-			Exception error;
-			BigDecimal value;
-			if (!TryParse(data, 0, data.Length, provider, out value, out error))
-				throw error;
-
-			if (context != null)
-				value.InplaceRound(context);
-
-			return value;
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a big integer instance. A fractional
-		 * part is discarded.
-		 *
-		 * @return this {@code BigDecimal} as a big integer instance.
-		 */
-
-		public BigInteger ToBigInteger() {
-			if ((_scale == 0) || (IsZero)) {
-				return GetUnscaledValue();
-			} else if (_scale < 0) {
-				return GetUnscaledValue().Multiply(Multiplication.PowerOf10(-(long) _scale));
-			} else {
-// (scale > 0)
-				return GetUnscaledValue().Divide(Multiplication.PowerOf10(_scale));
-			}
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a big integer instance if it has no
-		 * fractional part. If this {@code BigDecimal} has a fractional part, i.e.
-		 * if rounding would be necessary, an {@code ArithmeticException} is thrown.
-		 *
-		 * @return this {@code BigDecimal} as a big integer value.
-		 * @throws ArithmeticException
-		 *             if rounding is necessary.
-		 */
-
-		public BigInteger ToBigIntegerExact() {
-			if ((_scale == 0) || (IsZero)) {
-				return GetUnscaledValue();
-			} else if (_scale < 0) {
-				return GetUnscaledValue().Multiply(Multiplication.PowerOf10(-(long) _scale));
-			} else {
-// (scale > 0)
-				BigInteger integer;
-				BigInteger fraction;
-				// An optimization before do a heavy division
-				if ((_scale > AproxPrecision()) || (_scale > GetUnscaledValue().LowestSetBit)) {
-					// math.08=Rounding necessary
-					throw new ArithmeticException(Messages.math08); //$NON-NLS-1$
-				}
-				integer = GetUnscaledValue().DivideAndRemainder(Multiplication.PowerOf10(_scale), out fraction);
-				if (fraction.Sign != 0) {
-					// It exists a non-zero fractional part 
-					// math.08=Rounding necessary
-					throw new ArithmeticException(Messages.math08); //$NON-NLS-1$
-				}
-				return integer;
-			}
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as an long value. Any fractional part is
-		 * discarded. If the integral part of {@code this} is too big to be
-		 * represented as an long, then {@code this} % 2^64 is returned.
-		 *
-		 * @return this {@code BigDecimal} as a long value.
-		 */
-
-		public long ToInt64() {
-			/* If scale <= -64 there are at least 64 trailing bits zero in 10^(-scale).
-			 * If the scale is positive and very large the long value could be zero. */
-			return ((_scale <= -64) || (_scale > AproxPrecision()) ? 0L : ToBigInteger().ToInt64());
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a long value if it has no fractional
-		 * part and if its value fits to the int range ([-2^{63}..2^{63}-1]). If
-		 * these conditions are not met, an {@code ArithmeticException} is thrown.
-		 *
-		 * @return this {@code BigDecimal} as a long value.
-		 * @throws ArithmeticException
-		 *             if rounding is necessary or the number doesn't fit in a long.
-		 */
-
-		public long ToInt64Exact() {
-			return ValueExact(64);
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as an int value. Any fractional part is
-		 * discarded. If the integral part of {@code this} is too big to be
-		 * represented as an int, then {@code this} % 2^32 is returned.
-		 *
-		 * @return this {@code BigDecimal} as a int value.
-		 */
-
-		public int ToInt32() {
-			/* If scale <= -32 there are at least 32 trailing bits zero in 10^(-scale).
-			 * If the scale is positive and very large the long value could be zero. */
-			return ((_scale <= -32) || (_scale > AproxPrecision()) ? 0 : ToBigInteger().ToInt32());
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a int value if it has no fractional
-		 * part and if its value fits to the int range ([-2^{31}..2^{31}-1]). If
-		 * these conditions are not met, an {@code ArithmeticException} is thrown.
-		 *
-		 * @return this {@code BigDecimal} as a int value.
-		 * @throws ArithmeticException
-		 *             if rounding is necessary or the number doesn't fit in a int.
-		 */
-
-		public int ToInt32Exact() {
-			return (int) ValueExact(32);
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a short value if it has no fractional
-		 * part and if its value fits to the short range ([-2^{15}..2^{15}-1]). If
-		 * these conditions are not met, an {@code ArithmeticException} is thrown.
-		 *
-		 * @return this {@code BigDecimal} as a short value.
-		 * @throws ArithmeticException
-		 *             if rounding is necessary of the number doesn't fit in a
-		 *             short.
-		 */
-
-		public short ToInt16Exact() {
-			return (short) ValueExact(16);
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a byte value if it has no fractional
-		 * part and if its value fits to the byte range ([-128..127]). If these
-		 * conditions are not met, an {@code ArithmeticException} is thrown.
-		 *
-		 * @return this {@code BigDecimal} as a byte value.
-		 * @throws ArithmeticException
-		 *             if rounding is necessary or the number doesn't fit in a byte.
-		 */
-
-		public byte ToByteExact() {
-			return (byte) ValueExact(8);
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a float value. If {@code this} is too
-		 * big to be represented as an float, then {@code Float.POSITIVE_INFINITY}
-		 * or {@code Float.NEGATIVE_INFINITY} is returned.
-		 * <p>
-		 * Note, that if the unscaled value has more than 24 significant digits,
-		 * then this decimal cannot be represented exactly in a float variable. In
-		 * this case the result is rounded.
-		 * <p>
-		 * For example, if the instance {@code x1 = new BigDecimal("0.1")} cannot be
-		 * represented exactly as a float, and thus {@code x1.equals(new
-		 * BigDecimal(x1.folatValue())} returns {@code false} for this case.
-		 * <p>
-		 * Similarly, if the instance {@code new BigDecimal(16777217)} is converted
-		 * to a float, the result is {@code 1.6777216E}7.
-		 *
-		 * @return this {@code BigDecimal} as a float value.
-		 */
-
-		public float ToSingle() {
-			/* A similar code like in ToDouble() could be repeated here,
-			 * but this simple implementation is quite efficient. */
-			float floatResult = Sign;
-			long powerOfTwo = this._bitLength - (long) (_scale/Log10Of2);
-			if ((powerOfTwo < -149) || (floatResult == 0.0f)) {
-				// Cases which 'this' is very small
-				floatResult *= 0.0f;
-			} else if (powerOfTwo > 129) {
-				// Cases which 'this' is very large
-				floatResult *= Single.PositiveInfinity;
-			} else {
-				floatResult = (float) ToDouble();
-			}
-			return floatResult;
-		}
-
-		/**
-		 * Returns this {@code BigDecimal} as a double value. If {@code this} is too
-		 * big to be represented as an float, then {@code Double.POSITIVE_INFINITY}
-		 * or {@code Double.NEGATIVE_INFINITY} is returned.
-		 * <p>
-		 * Note, that if the unscaled value has more than 53 significant digits,
-		 * then this decimal cannot be represented exactly in a double variable. In
-		 * this case the result is rounded.
-		 * <p>
-		 * For example, if the instance {@code x1 = new BigDecimal("0.1")} cannot be
-		 * represented exactly as a double, and thus {@code x1.equals(new
-		 * BigDecimal(x1.ToDouble())} returns {@code false} for this case.
-		 * <p>
-		 * Similarly, if the instance {@code new BigDecimal(9007199254740993L)} is
-		 * converted to a double, the result is {@code 9.007199254740992E15}.
-		 * <p>
-		 *
-		 * @return this {@code BigDecimal} as a double value.
-		 */
-
-		public double ToDouble() {
-			int sign = Sign;
-			int exponent = 1076; // bias + 53
-			int lowestSetBit;
-			int discardedSize;
-			long powerOfTwo = this._bitLength - (long) (_scale/Log10Of2);
-			long bits; // IEEE-754 Standard
-			long tempBits; // for temporal calculations     
-			BigInteger mantisa;
-
-			if ((powerOfTwo < -1074) || (sign == 0)) {
-				// Cases which 'this' is very small            
-				return (sign*0.0d);
-			} else if (powerOfTwo > 1025) {
-				// Cases which 'this' is very large            
-				return (sign*Double.PositiveInfinity);
-			}
-			mantisa = GetUnscaledValue().Abs();
-			// Let be:  this = [u,s], with s > 0
-			if (_scale <= 0) {
-				// mantisa = abs(u) * 10^s
-				mantisa = mantisa.Multiply(Multiplication.PowerOf10(-_scale));
-			} else {
-				// (scale > 0)
-				BigInteger quotient;
-				BigInteger remainder;
-				BigInteger powerOfTen = Multiplication.PowerOf10(_scale);
-				int k = 100 - (int) powerOfTwo;
-				int compRem;
-
-				if (k > 0) {
-					/* Computing (mantisa * 2^k) , where 'k' is a enough big
-					 * power of '2' to can divide by 10^s */
-					mantisa = mantisa.ShiftLeft(k);
-					exponent -= k;
-				}
-				// Computing (mantisa * 2^k) / 10^s
-				quotient = mantisa.DivideAndRemainder(powerOfTen, out remainder);
-				// To check if the fractional part >= 0.5
-				compRem = remainder.ShiftLeftOneBit().CompareTo(powerOfTen);
-				// To add two rounded bits at end of mantisa
-				mantisa = quotient.ShiftLeft(2).Add(BigInteger.ValueOf((compRem*(compRem + 3))/2 + 1));
-				exponent -= 2;
-			}
-			lowestSetBit = mantisa.LowestSetBit;
-			discardedSize = mantisa.BitLength - 54;
-			if (discardedSize > 0) {
-				// (n > 54)
-				// mantisa = (abs(u) * 10^s) >> (n - 54)
-				bits = mantisa.ShiftRight(discardedSize).ToInt64();
-				tempBits = bits;
-				// #bits = 54, to check if the discarded fraction produces a carry             
-				if ((((bits & 1) == 1) && (lowestSetBit < discardedSize))
-				    || ((bits & 3) == 3)) {
-					bits += 2;
-				}
-			} else {
-				// (n <= 54)
-				// mantisa = (abs(u) * 10^s) << (54 - n)                
-				bits = mantisa.ToInt64() << -discardedSize;
-				tempBits = bits;
-				// #bits = 54, to check if the discarded fraction produces a carry:
-				if ((bits & 3) == 3) {
-					bits += 2;
-				}
-			}
-			// Testing bit 54 to check if the carry creates a new binary digit
-			if ((bits & 0x40000000000000L) == 0) {
-				// To drop the last bit of mantisa (first discarded)
-				bits >>= 1;
-				// exponent = 2^(s-n+53+bias)
-				exponent += discardedSize;
-			} else {
-				// #bits = 54
-				bits >>= 2;
-				exponent += discardedSize + 1;
-			}
-			// To test if the 53-bits number fits in 'double'            
-			if (exponent > 2046) {
-				// (exponent - bias > 1023)
-				return (sign*Double.PositiveInfinity);
-			}
-			if (exponent <= 0) {
-				// (exponent - bias <= -1023)
-				// Denormalized numbers (having exponent == 0)
-				if (exponent < -53) {
-					// exponent - bias < -1076
-					return (sign*0.0d);
-				}
-				// -1076 <= exponent - bias <= -1023 
-				// To discard '- exponent + 1' bits
-				bits = tempBits >> 1;
-				tempBits = bits & Utils.URShift(-1L, (63 + exponent));
-				bits >>= (-exponent);
-				// To test if after discard bits, a new carry is generated
-				if (((bits & 3) == 3) || (((bits & 1) == 1) && (tempBits != 0)
-				                          && (lowestSetBit < discardedSize))) {
-					bits += 1;
-				}
-				exponent = 0;
-				bits >>= 1;
-			}
-			// Construct the 64 double bits: [sign(1), exponent(11), mantisa(52)]
-			// bits = (long)((ulong)sign & 0x8000000000000000L) | ((long)exponent << 52) | (bits & 0xFFFFFFFFFFFFFL);
-			bits = sign & Int64.MinValue | ((long) exponent << 52) | (bits & 0xFFFFFFFFFFFFFL);
-			return BitConverter.Int64BitsToDouble(bits);
-		}
-
-		// TODO: must be verified
-		public decimal ToDecimal() {
-			var scaleDivisor = BigInteger.ValueOf(10).Pow(_scale);
-			var remainder = GetUnscaledValue().Remainder(scaleDivisor);
-			var scaledValue = GetUnscaledValue().Divide(scaleDivisor);
-
-			var leftOfDecimal = (decimal) scaledValue;
-			var rightOfDecimal = (remainder)/((decimal) scaleDivisor);
-
-			return leftOfDecimal + rightOfDecimal;
-		}
-
-		#endregion
-
-		#region Operators
-
-		public static BigDecimal operator +(BigDecimal a, BigDecimal b) {
-			// In case of implicit operators apply the precision of the dividend
-			return a.Add(b, new MathContext(a.Precision));
-		}
-
-		public static BigDecimal operator -(BigDecimal a, BigDecimal b) {
-			// In case of implicit operators apply the precision of the dividend
-			return a.Subtract(b, new MathContext(a.Precision));
-		}
-
-		public static BigDecimal operator /(BigDecimal a, BigDecimal b) {
-			// In case of implicit operators apply the precision of the dividend
-			return a.Divide(b, new MathContext(a.Precision));
-		}
-
-		public static BigDecimal operator %(BigDecimal a, BigDecimal b) {
-			// In case of implicit operators apply the precision of the dividend
-			return a.Remainder(b, new MathContext(a.Precision));
-		}
-
-		public static BigDecimal operator *(BigDecimal a, BigDecimal b) {
-			// In case of implicit operators apply the precision of the dividend
-			return a.Multiply(b, new MathContext(a.Precision));
-		}
-
-		public static BigDecimal operator +(BigDecimal a) {
-			return a.Plus();
-		}
-
-		public static BigDecimal operator -(BigDecimal a) {
-			return a.Negate();
-		}
-
-		public static bool operator ==(BigDecimal a, BigDecimal b) {
-			if ((object) a == null && (object) b == null)
-				return true;
-			if ((object) a == null)
-				return false;
-			return a.Equals(b);
-		}
-
-		public static bool operator !=(BigDecimal a, BigDecimal b) {
-			return !(a == b);
-		}
-
-		public static bool operator >(BigDecimal a, BigDecimal b) {
-			return a.CompareTo(b) < 0;
-		}
-
-		public static bool operator <(BigDecimal a, BigDecimal b) {
-			return a.CompareTo(b) > 0;
-		}
-
-		public static bool operator >=(BigDecimal a, BigDecimal b) {
-			return a == b || a > b;
-		}
-
-		public static bool operator <=(BigDecimal a, BigDecimal b) {
-			return a == b || a < b;
-		}
-
-		#endregion
-
-		#region Implicit Operators
-
-		public static implicit operator Int16(BigDecimal d) {
-			return d.ToInt16Exact();
-		}
-
-		public static implicit operator Int32(BigDecimal d) {
-			return d.ToInt32();
-		}
-
-		public static implicit operator Int64(BigDecimal d) {
-			return d.ToInt64();
-		}
-
-		public static implicit operator Single(BigDecimal d) {
-			return d.ToSingle();
-		}
-
-		public static implicit operator Double(BigDecimal d) {
-			return d.ToDouble();
-		}
-
-		public static implicit operator BigInteger(BigDecimal d) {
-			return d.ToBigInteger();
-		}
-
-		public static implicit operator String(BigDecimal d) {
-			return d.ToString();
-		}
-
-		public static implicit operator BigDecimal(long value) {
-			return new BigDecimal(value);
-		}
-
-		public static implicit operator BigDecimal(double value) {
-			return new BigDecimal(value);
-		}
-
-		public static implicit operator BigDecimal(int value) {
-			return new BigDecimal(value);
-		}
-
-		public static implicit operator BigDecimal(BigInteger value) {
-			return new BigDecimal(value);
-		}
-
-		#endregion
 	}
 }
